// types/fingerprint.ts
export interface WebGLParams {
  p1: string;
  p2: string;
  p3: string;
  p4: string;
  p5: string;
  p6: string;
  p7: string;
  p8: string;
  p9: string;
  p10: string;
  p11: string;
  p12: string;
  p13: string;
  p14: string;
  p15: string;
  p16: string;
  p17: string;
  p18: string;
}

export interface Fingerprint {
  c: string;
  wg: string | WebGLParams;
  h: string;
}